requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
cycler==0.11.0
decouple==0.0.7
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==4.0.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-extensions==3.1.5
django-heroku==0.3.1
django-on-heroku==1.1.2
django-rest-knox==4.2.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
environ==1.0
Flask==2.1.2
fonttools==4.34.4
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
mysqlclient==2.1.0
numpy==1.22.4
packaging==21.3
pandas==1.4.2
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
Pygments==2.12.0
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
scipy==1.8.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
Werkzeug==2.1.2
whitenoise==6.2.0
